+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.6
+2016/11/12
+- Added delete action
+- Added matching progress
+
 # v0.5
 2016/10/28
 - Added master and copy counts per path